+++ lua/mini/indentscope.lua
@@ -171,6 +171,12 @@ end
 ---        1 and 2    2-5 | 2-4
 ---      3 and more   2-4 | 2-4
 --- <
+--- - Option `n_lines` defines |MiniIndentscope.get_scope()| behavior for how many
+---   lines above/below to check before iteration is stopped. Scope that reached
+---   computation limit has <is_incomplete> field set to `true`.
+---   Lower values will result in better overall performance in exchange for more
+---   frequent incomplete scope computation. Set to `math.huge` for no restriction.
+---
 --- - Option `try_as_border` controls how to act when input line can be
 ---   recognized as a border of some neighbor indent scope. In main example,
 ---   when input line is 1 and can be recognized as border for inner scope,
@@ -216,6 +222,9 @@ MiniIndentscope.config = {
     -- Useful to see incremental scopes with horizontal cursor movements.
     indent_at_cursor = true,
 
+    -- Maximum number of lines above or below within which scope is computed
+    n_lines = 10000,
+
     -- Whether to first check input line to be a border of adjacent scope.
     -- Use it if you want to place cursor on function header to get scope of
     -- its body.
@@ -249,12 +258,13 @@ MiniIndentscope.config = {
 --- - Compute reference "indent at column". Reference line is an input `line`
 ---   which might be modified to one of its neighbors if `try_as_border` option
 ---   is `true`: if it can be viewed as border of some neighbor scope, it will.
---- - Process upwards and downwards from reference line to search for line with
+--- - Process upwards and downwards from reference line searching for line with
 ---   indent strictly less than reference one. This is like casting rays up and
 ---   down from reference line and reference indent until meeting "a wall"
 ---   (character to the right of indent or buffer edge). Latest line before
----   meeting is a respective end of scope body. It always exists because
+---   meeting a wall is a respective end of scope body. It always exists because
 ---   reference line is a such one.
+---   Casting ray is forced to stop if it goes over `opts.n_lines` lines.
 --- - Based on top and bottom lines with strictly lower indent, construct
 ---   scopes's border. The way it is computed is decided based on `border`
 ---   option (see |MiniIndentscope.config| for more information).
@@ -285,7 +295,8 @@ MiniIndentscope.config = {
 ---@return table Table with scope information:
 ---   - <body> - table with <top> (top line of scope, inclusive), <bottom>
 ---     (bottom line of scope, inclusive), and <indent> (minimum indent within
----     scope) keys. Line numbers start at 1.
+---     scope) keys. Line numbers start at 1. Can also have <is_incomplete> key
+---     set to `true` if computation was stopped due to `opts.n_lines` restriction.
 ---   - <border> - table with <top> (line of top border, might be `nil`),
 ---     <bottom> (line of bottom border, might be `nil`), and <indent> (indent
 ---     of border) keys. Line numbers start at 1.
@@ -316,10 +327,11 @@ MiniIndentscope.get_scope = function(line, col, opts)
   if indent <= 0 then
     body.top, body.bottom, body.indent = 1, vim.fn.line('$'), line_indent
   else
-    local up_min_indent, down_min_indent
-    body.top, up_min_indent = H.cast_ray(line, indent, 'up', opts)
-    body.bottom, down_min_indent = H.cast_ray(line, indent, 'down', opts)
+    local up_line, up_min_indent, up_is_incomplete = H.cast_ray(line, indent, 'up', opts)
+    local down_line, down_min_indent, down_is_incomplete = H.cast_ray(line, indent, 'down', opts)
+    body.top, body.bottom = up_line, down_line
     body.indent = math.min(line_indent, up_min_indent, down_min_indent)
+    body.is_incomplete = up_is_incomplete or down_is_incomplete
   end
 
   return {
@@ -631,6 +643,7 @@ H.setup_config = function(config)
 
     ['options.border'] = { config.options.border, 'string' },
     ['options.indent_at_cursor'] = { config.options.indent_at_cursor, 'boolean' },
+    ['options.n_lines'] = { config.options.n_lines, 'number' },
     ['options.try_as_border'] = { config.options.try_as_border, 'boolean' },
   })
   return config
@@ -731,10 +744,11 @@ H.get_line_indent = function(line, opts)
   local prev_nonblank = vim.fn.prevnonblank(line)
   local res = vim.fn.indent(prev_nonblank)
 
-  -- Compute indent of blank line depending on `options.border` values
+  -- Compute indent of blank line depending on `border` option
+  local border = opts.border
   if line ~= prev_nonblank then
     local next_indent = vim.fn.indent(vim.fn.nextnonblank(line))
-    local blank_rule = H.blank_indent_funs[opts.border]
+    local blank_rule = H.blank_indent_funs[border]
     res = blank_rule(res, next_indent)
   end
 
@@ -747,14 +761,19 @@ H.cast_ray = function(line, indent, direction, opts)
     final_line, increment = vim.fn.line('$'), 1
   end
 
+  local is_incomplete
+  if math.abs(line - final_line) > opts.n_lines then
+    final_line, is_incomplete = line + increment * opts.n_lines, true
+  end
+
   local min_indent = math.huge
   for l = line, final_line, increment do
     local new_indent = H.get_line_indent(l + increment, opts)
-    if new_indent < indent then return l, min_indent end
+    if new_indent < indent then return l, min_indent, nil end
     if new_indent < min_indent then min_indent = new_indent end
   end
 
-  return final_line, min_indent
+  return final_line, min_indent, is_incomplete
 end
 
 H.scope_get_draw_indent = function(scope) return scope.border.indent or (scope.body.indent - 1) end